requirements.txt
haystack-ai
duckduckgo-api-haystack
transformers
sentence-transformers
datasets
apify-haystack
fastapi
uvicorn
openai